Move ImmutableList<T>.Enumerator to a new file
