+++ src/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -71,6 +71,7 @@
     <Compile Include="System\Collections\Immutable\ImmutableList.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableList_1.Builder.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableList_1.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableList_1.Enumerator.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableList_1.Node.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableQueue.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableQueue_1.cs" />
